Continued work on EK80 support.
